+++ components/import/steps/id-integration/index.ts
@@ -1,0 +1,12 @@
+/**
+ * ID統合ステップのコンポーネント
+ */
+
+export { ClassIntegrationPanel } from "./ClassIntegrationPanel"
+export { DetailPanel } from "./DetailPanel"
+export { MatchedItemRow } from "./MatchedItemRow"
+export { NoMatchItemRow } from "./NoMatchItemRow"
+export { StrategyOption } from "./StrategyOption"
+export { StudentIntegrationPanel } from "./StudentIntegrationPanel"
+export { SubtotalGroupIntegrationPanel } from "./SubtotalGroupIntegrationPanel"
+export * from "./types"
