+++ electron-src/lib/import/merge/matchers/index.ts
@@ -1,0 +1,12 @@
+/**
+ * マッチャーモジュール
+ */
+
+export { matchClasses, preMatchClasses } from "./classMatcher"
+export { matchStudents, preMatchStudents } from "./studentMatcher"
+export {
+  matchSubtotalGroups,
+  preMatchSubtotalGroups,
+} from "./subtotalGroupMatcher"
+export * from "./types"
+export { matchUsers } from "./userMatcher"
