+++ electron-src/lib/import/merge/processors/index.ts
@@ -1,0 +1,7 @@
+/**
+ * ID統合処理のプロセッサ
+ */
+
+export { processClassIdIntegration } from "./classProcessor"
+export { processStudentIdIntegration } from "./studentProcessor"
+export { processSubtotalGroupIdIntegration } from "./subtotalGroupProcessor"
