+++ electron-src/lib/import/merge/processors/studentProcessor.ts
@@ -1,0 +1,186 @@
+/**
+ * 生徒のID統合処理
+ */
+
+import { randomUUID } from "crypto"
+
+import type {
+  FileOverviewData,
+  IdIntegrationConfig,
+  IdIntegrationDecision,
+} from "../../../../../types/projectArchive.types"
+import type { ExtractedArchiveData } from "../../project-archive/archiveExtractor"
+import type {
+  IdChangeTarget,
+  IdMappings,
+  ImportCounts,
+  PrismaTransaction,
+} from "../types"
+
+/**
+ * 生徒のID統合処理を実行
+ */
+export async function processStudentIdIntegration(
+  data: ExtractedArchiveData,
+  preMatchResult: FileOverviewData,
+  config: IdIntegrationConfig["student"],
+  idMappings: IdMappings,
+  idChangeTargets: IdChangeTarget[],
+  counts: ImportCounts,
+  warnings: string[],
+  tx: PrismaTransaction
+): Promise<void> {
+  const studentPreMatch = preMatchResult.student
+
+  // ID一致したもの（自動で紐づく）
+  for (const match of studentPreMatch.byId) {
+    idMappings.student[match.importId] = match.existingId
+  }
+
+  // ID不一致のものを処理
+  const processDecision = async (
+    importId: string,
+    decision: IdIntegrationDecision | undefined,
+    defaultExistingId: string | undefined
+  ) => {
+    const importStudent = data.studentsData.students.find(
+      (s) => s.id === importId
+    )
+    if (!importStudent) return
+
+    if (!decision || decision.decisionType === "create_new") {
+      // 新規作成
+      const existingByStudentNumber = await tx.student.findUnique({
+        where: { studentNumber: importStudent.studentNumber },
+      })
+
+      if (existingByStudentNumber) {
+        idMappings.student[importId] = existingByStudentNumber.id
+        warnings.push(
+          `生徒「${importStudent.lastName} ${importStudent.firstName}」は既存データを使用します`
+        )
+      } else {
+        const newId = randomUUID()
+        await tx.student.create({
+          data: {
+            id: newId,
+            studentNumber: importStudent.studentNumber,
+            lastName: importStudent.lastName,
+            firstName: importStudent.firstName,
+            lastNameKana: importStudent.lastNameKana,
+            firstNameKana: importStudent.firstNameKana,
+            enrollmentYear: importStudent.enrollmentYear,
+          },
+        })
+        idMappings.student[importId] = newId
+        counts.created.students++
+      }
+    } else if (decision.decisionType === "same_person") {
+      const existingId = decision.existingId || defaultExistingId
+      if (!existingId) {
+        warnings.push(
+          `生徒「${importStudent.lastName} ${importStudent.firstName}」の既存IDが見つかりません`
+        )
+        return
+      }
+
+      idMappings.student[importId] = existingId
+
+      if (decision.idChoice === "use_import_id") {
+        // Stage 2でID変更を行う
+        idChangeTargets.push({
+          category: "student",
+          existingId: existingId,
+          newId: importId,
+        })
+      }
+    } else if (decision.decisionType === "skip") {
+      counts.skipped.students++
+    }
+  }
+
+  // 学籍番号一致
+  if (studentPreMatch.byStudentNumber) {
+    for (const match of studentPreMatch.byStudentNumber) {
+      const decision = config.decisions.find(
+        (d) => d.importId === match.importId
+      )
+
+      // strategyに応じたデフォルト処理
+      if (config.strategy === "by_student_number") {
+        await processDecision(
+          match.importId,
+          decision || {
+            importId: match.importId,
+            decisionType: "same_person",
+            existingId: match.existingId,
+            idChoice: "use_existing_id",
+          },
+          match.existingId
+        )
+      } else if (config.strategy === "all_new") {
+        await processDecision(
+          match.importId,
+          decision || {
+            importId: match.importId,
+            decisionType: "create_new",
+          },
+          undefined
+        )
+      } else {
+        await processDecision(match.importId, decision, match.existingId)
+      }
+    }
+  }
+
+  // 氏名一致
+  if (studentPreMatch.byName) {
+    for (const match of studentPreMatch.byName) {
+      // byStudentNumberで既に処理済みの場合はスキップ
+      if (idMappings.student[match.importId]) continue
+
+      const decision = config.decisions.find(
+        (d) => d.importId === match.importId
+      )
+
+      if (config.strategy === "by_name") {
+        await processDecision(
+          match.importId,
+          decision || {
+            importId: match.importId,
+            decisionType: "same_person",
+            existingId: match.existingId,
+            idChoice: "use_existing_id",
+          },
+          match.existingId
+        )
+      } else if (config.strategy === "all_new") {
+        await processDecision(
+          match.importId,
+          decision || {
+            importId: match.importId,
+            decisionType: "create_new",
+          },
+          undefined
+        )
+      } else {
+        await processDecision(match.importId, decision, match.existingId)
+      }
+    }
+  }
+
+  // どれにも一致しない
+  for (const item of studentPreMatch.noMatch) {
+    if (idMappings.student[item.importId]) continue
+
+    const decision = config.decisions.find((d) => d.importId === item.importId)
+    await processDecision(
+      item.importId,
+      decision || {
+        importId: item.importId,
+        decisionType: "create_new",
+      },
+      undefined
+    )
+  }
+}
