+++ electron-src/lib/import/project-archive/dataCreator.ts
@@ -4,92 +4,16 @@
  * インポートデータをデータベースに作成
  */
 
-import type { PrismaClient } from "@prisma/client"
-import * as fs from "fs"
-import * as path from "path"
-
 import type { ArchiveDataCounts } from "../../../../types/projectArchive.types"
-import { getDataDirectory } from "../../dataManager"
 import prisma from "../../prisma/client"
 import type { ExtractedArchiveData } from "./archiveExtractor"
 import type { IdMappings } from "./idRemapper"
 import { remapId, remapIdRequired } from "./idRemapper"
-
-/** Prismaトランザクションクライアント型 */
-type TransactionClient = Omit<
-  PrismaClient,
-  "$connect" | "$disconnect" | "$on" | "$transaction" | "$use" | "$extends"
->
-
-/**
- * 重複しないstudentNumberを生成
- *
- * 既存のstudentNumberがある場合は `_1`, `_2` のようなサフィックスを付与して
- * 一意性を保証する
- *
- * @param tx - Prismaトランザクションクライアント
- * @param originalStudentNumber - 元のstudentNumber
- * @returns 一意なstudentNumber
- */
-async function generateUniqueStudentNumber(
-  tx: TransactionClient,
-  originalStudentNumber: string
-): Promise<string> {
-  const existing = await tx.student.findUnique({
-    where: { studentNumber: originalStudentNumber },
-  })
-
-  if (!existing) {
-    return originalStudentNumber
-  }
-
-  // サフィックスを付けて重複を回避
-  let suffix = 1
-  let newStudentNumber = `${originalStudentNumber}_${suffix}`
-
-  while (
-    await tx.student.findUnique({ where: { studentNumber: newStudentNumber } })
-  ) {
-    suffix++
-    newStudentNumber = `${originalStudentNumber}_${suffix}`
-  }
-
-  return newStudentNumber
-}
-
-/**
- * 重複しない学級名を生成
- *
- * 既存の名前がある場合は `(2)`, `(3)` のようなサフィックスを付与して
- * 一意性を保証する
- *
- * @param tx - Prismaトランザクションクライアント
- * @param originalName - 元の学級名
- * @returns 一意な学級名
- */
-async function generateUniqueClassName(
-  tx: TransactionClient,
-  originalName: string
-): Promise<string> {
-  const existing = await tx.class.findUnique({
-    where: { name: originalName },
-  })
-
-  if (!existing) {
-    return originalName
-  }
-
-  // サフィックスを付けて重複を回避
-  let suffix = 2
-  let newName = `${originalName} (${suffix})`
-
-  while (await tx.class.findUnique({ where: { name: newName } })) {
-    suffix++
-    newName = `${originalName} (${suffix})`
-  }
-
-  return newName
-}
+import { copyImages, createImageRecords } from "./imageHandler"
+import {
+  generateUniqueClassName,
+  generateUniqueStudentNumber,
+} from "./uniqueNameGenerators"
 
 /**
  * データ作成結果
@@ -437,156 +361,6 @@ export async function createImportedData(
       success: false,
       error:
         error instanceof Error ? error.message : "データの作成に失敗しました",
-    }
-  }
-}
-
-/**
- * 画像ファイルをプロジェクトディレクトリにコピー
- *
- * @param data - 展開されたアーカイブデータ
- * @param newProjectId - 新規プロジェクトID
- */
-async function copyImages(
-  data: ExtractedArchiveData,
-  newProjectId: string
-): Promise<void> {
-  const dataDir = getDataDirectory()
-  const projectDir = path.join(dataDir, "projects", newProjectId)
-
-  // プロジェクトディレクトリを作成
-  const masterImagesDir = path.join(projectDir, "master-images")
-  const answerSheetsDir = path.join(projectDir, "answer-sheets")
-  fs.mkdirSync(masterImagesDir, { recursive: true })
-  fs.mkdirSync(answerSheetsDir, { recursive: true })
-
-  // マスター画像をコピー
-  for (const srcPath of data.masterImagePaths) {
-    const filename = path.basename(srcPath)
-    const destPath = path.join(masterImagesDir, filename)
-    fs.copyFileSync(srcPath, destPath)
-  }
-
-  // 答案画像をコピー（生徒IDでサブディレクトリを分ける場合がある）
-  for (const srcPath of data.answerSheetPaths) {
-    // 相対パス構造を維持
-    const relativePath = srcPath.substring(
-      srcPath.indexOf("answer-sheets") + "answer-sheets".length + 1
-    )
-    const destPath = path.join(answerSheetsDir, relativePath)
-
-    // ディレクトリを作成
-    fs.mkdirSync(path.dirname(destPath), { recursive: true })
-    fs.copyFileSync(srcPath, destPath)
-  }
-}
-
-/**
- * 画像レコードを作成（MasterImage / StudentAnswerImage）
- *
- * - v1.2.0+: masterImages と studentAnswerImages を使用
- * - v1.1.0以前: pageImages から変換（後方互換性）
- *
- * @param data - 展開されたアーカイブデータ
- * @param mappings - IDマッピング
- * @param newProjectId - 新規プロジェクトID
- */
-async function createImageRecords(
-  data: ExtractedArchiveData,
-  mappings: IdMappings,
-  newProjectId: string
-): Promise<void> {
-  // v1.2.0+ 形式: masterImages と studentAnswerImages が存在する場合
-  if (
-    data.projectData.masterImages &&
-    data.projectData.masterImages.length > 0
-  ) {
-    for (const img of data.projectData.masterImages) {
-      const newProjectPageId = remapId(img.projectPageId, mappings.projectPage)
-      if (!newProjectPageId) continue
-
-      const filename = path.basename(img.imagePath)
-      const newImagePath = `projects/${newProjectId}/master-images/${filename}`
-
-      await prisma.masterImage.create({
-        data: {
-          id: remapIdRequired(img.id, mappings.masterImage),
-          projectPageId: newProjectPageId,
-          imagePath: newImagePath,
-        },
-      })
-    }
-  }
-
-  if (
-    data.projectData.studentAnswerImages &&
-    data.projectData.studentAnswerImages.length > 0
-  ) {
-    for (const img of data.projectData.studentAnswerImages) {
-      const newProjectPageId = remapId(img.projectPageId, mappings.projectPage)
-      const newStudentId = remapId(img.studentId, mappings.student)
-      if (!newProjectPageId || !newStudentId) continue
-
-      const relativePath = img.imagePath.substring(
-        img.imagePath.indexOf("answer-sheets") + "answer-sheets".length + 1
-      )
-      const newImagePath =
-        `projects/${newProjectId}/answer-sheets/${relativePath}`.replace(
-          /\\/g,
-          "/"
-        )
-
-      await prisma.studentAnswerImage.create({
-        data: {
-          id: remapIdRequired(img.id, mappings.studentAnswerImage),
-          projectPageId: newProjectPageId,
-          studentId: newStudentId,
-          imagePath: newImagePath,
-        },
-      })
-    }
-
-    return
-  }
-
-  // v1.1.0以前: pageImages から変換（後方互換性）
-  for (const img of data.projectData.pageImages) {
-    const newProjectPageId = remapId(img.projectPageId, mappings.projectPage)
-    if (!newProjectPageId) continue
-
-    const filename = path.basename(img.imagePath)
-
-    if (img.imageType === "MODEL_ANSWER") {
-      const newImagePath = `projects/${newProjectId}/master-images/${filename}`
-
-      await prisma.masterImage.create({
-        data: {
-          id: remapIdRequired(img.id, mappings.pageImage),
-          projectPageId: newProjectPageId,
-          imagePath: newImagePath,
-        },
-      })
-    } else if (img.imageType === "STUDENT_ANSWER" && img.studentId) {
-      const newStudentId = remapId(img.studentId, mappings.student)
-      if (!newStudentId) continue
-
-      const relativePath = img.imagePath.substring(
-        img.imagePath.indexOf("answer-sheets") + "answer-sheets".length + 1
-      )
-      const newImagePath =
-        `projects/${newProjectId}/answer-sheets/${relativePath}`.replace(
-          /\\/g,
-          "/"
-        )
-
-      await prisma.studentAnswerImage.create({
-        data: {
-          id: remapIdRequired(img.id, mappings.pageImage),
-          projectPageId: newProjectPageId,
-          studentId: newStudentId,
-          imagePath: newImagePath,
-        },
-      })
     }
   }
 }
