+++ electron-src/lib/import/project-archive/imageHandler.ts
@@ -1,0 +1,162 @@
+/**
+ * 新規インポート用の画像処理モジュール
+ */
+
+import * as fs from "fs"
+import * as path from "path"
+
+import { getDataDirectory } from "../../dataManager"
+import prisma from "../../prisma/client"
+import type { ExtractedArchiveData } from "./archiveExtractor"
+import type { IdMappings } from "./idRemapper"
+import { remapId, remapIdRequired } from "./idRemapper"
+
+/**
+ * 画像ファイルをプロジェクトディレクトリにコピー
+ *
+ * @param data - 展開されたアーカイブデータ
+ * @param newProjectId - 新規プロジェクトID
+ */
+export async function copyImages(
+  data: ExtractedArchiveData,
+  newProjectId: string
+): Promise<void> {
+  const dataDir = getDataDirectory()
+  const projectDir = path.join(dataDir, "projects", newProjectId)
+
+  // プロジェクトディレクトリを作成
+  const masterImagesDir = path.join(projectDir, "master-images")
+  const answerSheetsDir = path.join(projectDir, "answer-sheets")
+  fs.mkdirSync(masterImagesDir, { recursive: true })
+  fs.mkdirSync(answerSheetsDir, { recursive: true })
+
+  // マスター画像をコピー
+  for (const srcPath of data.masterImagePaths) {
+    const filename = path.basename(srcPath)
+    const destPath = path.join(masterImagesDir, filename)
+    fs.copyFileSync(srcPath, destPath)
+  }
+
+  // 答案画像をコピー（生徒IDでサブディレクトリを分ける場合がある）
+  for (const srcPath of data.answerSheetPaths) {
+    // 相対パス構造を維持
+    const relativePath = srcPath.substring(
+      srcPath.indexOf("answer-sheets") + "answer-sheets".length + 1
+    )
+    const destPath = path.join(answerSheetsDir, relativePath)
+
+    // ディレクトリを作成
+    fs.mkdirSync(path.dirname(destPath), { recursive: true })
+    fs.copyFileSync(srcPath, destPath)
+  }
+}
+
+/**
+ * 画像レコードを作成（MasterImage / StudentAnswerImage）
+ *
+ * - v1.2.0+: masterImages と studentAnswerImages を使用
+ * - v1.1.0以前: pageImages から変換（後方互換性）
+ *
+ * @param data - 展開されたアーカイブデータ
+ * @param mappings - IDマッピング
+ * @param newProjectId - 新規プロジェクトID
+ */
+export async function createImageRecords(
+  data: ExtractedArchiveData,
+  mappings: IdMappings,
+  newProjectId: string
+): Promise<void> {
+  // v1.2.0+ 形式: masterImages と studentAnswerImages が存在する場合
+  if (
+    data.projectData.masterImages &&
+    data.projectData.masterImages.length > 0
+  ) {
+    for (const img of data.projectData.masterImages) {
+      const newProjectPageId = remapId(img.projectPageId, mappings.projectPage)
+      if (!newProjectPageId) continue
+
+      const filename = path.basename(img.imagePath)
+      const newImagePath = `projects/${newProjectId}/master-images/${filename}`
+
+      await prisma.masterImage.create({
+        data: {
+          id: remapIdRequired(img.id, mappings.masterImage),
+          projectPageId: newProjectPageId,
+          imagePath: newImagePath,
+        },
+      })
+    }
+  }
+
+  if (
+    data.projectData.studentAnswerImages &&
+    data.projectData.studentAnswerImages.length > 0
+  ) {
+    for (const img of data.projectData.studentAnswerImages) {
+      const newProjectPageId = remapId(img.projectPageId, mappings.projectPage)
+      const newStudentId = remapId(img.studentId, mappings.student)
+      if (!newProjectPageId || !newStudentId) continue
+
+      const relativePath = img.imagePath.substring(
+        img.imagePath.indexOf("answer-sheets") + "answer-sheets".length + 1
+      )
+      const newImagePath =
+        `projects/${newProjectId}/answer-sheets/${relativePath}`.replace(
+          /\\/g,
+          "/"
+        )
+
+      await prisma.studentAnswerImage.create({
+        data: {
+          id: remapIdRequired(img.id, mappings.studentAnswerImage),
+          projectPageId: newProjectPageId,
+          studentId: newStudentId,
+          imagePath: newImagePath,
+        },
+      })
+    }
+
+    return
+  }
+
+  // v1.1.0以前: pageImages から変換（後方互換性）
+  for (const img of data.projectData.pageImages) {
+    const newProjectPageId = remapId(img.projectPageId, mappings.projectPage)
+    if (!newProjectPageId) continue
+
+    const filename = path.basename(img.imagePath)
+
+    if (img.imageType === "MODEL_ANSWER") {
+      const newImagePath = `projects/${newProjectId}/master-images/${filename}`
+
+      await prisma.masterImage.create({
+        data: {
+          id: remapIdRequired(img.id, mappings.pageImage),
+          projectPageId: newProjectPageId,
+          imagePath: newImagePath,
+        },
+      })
+    } else if (img.imageType === "STUDENT_ANSWER" && img.studentId) {
+      const newStudentId = remapId(img.studentId, mappings.student)
+      if (!newStudentId) continue
+
+      const relativePath = img.imagePath.substring(
+        img.imagePath.indexOf("answer-sheets") + "answer-sheets".length + 1
+      )
+      const newImagePath =
+        `projects/${newProjectId}/answer-sheets/${relativePath}`.replace(
+          /\\/g,
+          "/"
+        )
+
+      await prisma.studentAnswerImage.create({
+        data: {
+          id: remapIdRequired(img.id, mappings.pageImage),
+          projectPageId: newProjectPageId,
+          studentId: newStudentId,
+          imagePath: newImagePath,
+        },
+      })
+    }
+  }
+}
