+++ hooks/import/constants.ts
@@ -1,0 +1,53 @@
+/**
+ * インポートウィザードの定数
+ */
+
+import type {
+  ImportWizardState,
+  ImportWizardStep,
+} from "@/types/projectArchive.types"
+
+/**
+ * インポートウィザードの初期状態
+ */
+export const initialState: ImportWizardState = {
+  currentStep: "file_select",
+  archivePath: null,
+  manifest: null,
+  fileOverviewData: null,
+  idIntegrationConfig: {
+    student: { strategy: "by_student_number", decisions: [] },
+    class: { strategy: "by_name", decisions: [] },
+    subtotalGroup: { strategy: "by_name", decisions: [] },
+  },
+  scoringConflictConfig: {
+    strategy: "newer_wins",
+    manualResolutions: {},
+  },
+  matchingConfig: {
+    student: "studentNumber",
+    class: "name",
+    user: "username",
+    project: "always_new",
+    subtotalGroup: "name",
+  },
+  isProcessing: false,
+  error: null,
+  matchingSummaries: [],
+  matchingDecisions: {},
+  updateDecisions: {},
+}
+
+/**
+ * ステップの順序
+ * file_select → file_overview → id_integration → scoring_conflict → update_confirm → final_confirm → execute
+ */
+export const STEP_ORDER: ImportWizardStep[] = [
+  "file_select",
+  "file_overview",
+  "id_integration",
+  "scoring_conflict",
+  "update_confirm",
+  "final_confirm",
+  "execute",
+]
