+++ hooks/import/useImportWizard.ts
@@ -10,57 +10,12 @@ import type {
   IdIntegrationConfig,
   IdIntegrationDecision,
   ImportWizardState,
-  ImportWizardStep,
   MatchingConfig,
   MatchingDecisionType,
-  ScoringConflictConfig,
   ScoringConflictResolutionStrategy,
 } from "@/types/projectArchive.types"
 
-/**
- * インポートウィザードの初期状態
- */
-const initialState: ImportWizardState = {
-  currentStep: "file_select",
-  archivePath: null,
-  manifest: null,
-  fileOverviewData: null,
-  idIntegrationConfig: {
-    student: { strategy: "by_student_number", decisions: [] },
-    class: { strategy: "by_name", decisions: [] },
-    subtotalGroup: { strategy: "by_name", decisions: [] },
-  },
-  scoringConflictConfig: {
-    strategy: "newer_wins",
-    manualResolutions: {},
-  },
-  matchingConfig: {
-    student: "studentNumber",
-    class: "name",
-    user: "username",
-    project: "always_new",
-    subtotalGroup: "name",
-  },
-  isProcessing: false,
-  error: null,
-  matchingSummaries: [],
-  matchingDecisions: {},
-  updateDecisions: {},
-}
-
-/**
- * ステップの順序
- * file_select → file_overview → id_integration → scoring_conflict → update_confirm → final_confirm → execute
- */
-const STEP_ORDER: ImportWizardStep[] = [
-  "file_select",
-  "file_overview",
-  "id_integration",
-  "scoring_conflict",
-  "update_confirm",
-  "final_confirm",
-  "execute",
-]
+import { initialState, STEP_ORDER } from "./constants"
 
 /**
  * インポートウィザードの状態管理フック
@@ -403,7 +358,9 @@ export function useImportWizard() {
   const setAllScoringConflictResolutions = useCallback(
     (conflictIds: string[], resolution: "import" | "existing") => {
       setState((prev) => {
-        const newResolutions = { ...prev.scoringConflictConfig.manualResolutions }
+        const newResolutions = {
+          ...prev.scoringConflictConfig.manualResolutions,
+        }
         for (const id of conflictIds) {
           newResolutions[id] = resolution
         }
